requirements.txt
altgraph==0.17.4
appdirs==1.4.4
argcomplete==3.6.2
blinker==1.9.0
click==8.1.8
coverage==7.8.0
Flask==3.1.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pytest==8.3.5
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.1
setuptools==79.0.0
six==1.17.0
Werkzeug==3.1.3
wheel==0.45.1
